Add particle effects for laser collisions
 * Refactored game entities into a module
 * Fix bug in raytracing collision detection where the edge from the past point to the first one wasn't being tested for collisions
 * Re-activate production optimizations with webpack and switch from optimizing for size to optimizing for performance (`-Oz` to `-O4`)
 * Removed obnoxious circle animations on mouse move

+++ frontend/game-engine/src/lib.rs
@@ -51,8 +51,8 @@ extern "C" {
         y: u16,
         width: u16,
         radius: u16,
-        startAngle: f64,
-        endAngle: f64,
+        startAngle: f32,
+        endAngle: f32,
         counterClockwise: bool,
     );
     pub fn render_line(r: u8, g: u8, b: u8, width: u16, x1: u16, y1: u16, x2: u16, y2: u16);
